Open source httperf2
Summary:
This is a tool we've used for performance benchmarking for many years.  It sends requests for the same URL to the same server repeatedly, with many command line flags to control behavior.

It was modeled around the idea of a "Client" that would establish an HTTP connection and send a number of requests on that connection, then close.

Reviewed By: lnicco

Differential Revision: D41364808

fbshipit-source-id: f96526b548b33f176a5e34ede41d18280d4dd72a