Add minimal Rust port of Measure module
Summary: Add a Rust version of the Measure module which doesn't implement distribution bucketing (which we don't seem to use) or serialization/deserialization. Integrate with the OCaml implementation: when pushing to the global record stack, also push to Rust's stack. When popping, also pop from Rust's stack. Since each MultiWorker worker invokes a push when starting the job and a pop when completing, this causes them to report back the Rust results unioned with the OCaml.

Reviewed By: shayne-fletcher

Differential Revision: D41361426

fbshipit-source-id: b178a32c7aa4f81add12bbde57b47cfeb7e954ed