fanout tests: test that expected fanout is included in actual fanout instead of testing for equality
Summary:
High-level goal: make the fanout tests easier and faster to run

To that end, I would like to write something like an hh_single_fanout binary which encapsulate the fanout logic and can be run in batch mode by verify.py

Ideally, the current slow fanout tests could be run only in the CI. For that to be possible, all fanout tests should use the same .exp files.

This diff allows to unify all the .exp files.

To do that, we need to check inclusion, rather than equality, of the expected fanout in the actual fanout.

A consequence of this diff is that we won't check for errors before and after anymore.

This diff introduces an option `--check-expected-included-in-actual` to verify.py to that end.

The interesting files to review are:
```
M hphp/hack/test/fanout/TARGETS
M hphp/hack/test/fanout/tools/fanout_test_driver.py
M hphp/hack/test/verify.py
```
(Use the file filter box to show them)

Reviewed By: bobrenjc93

Differential Revision: D41470423

fbshipit-source-id: ddf4e3464c1a46941cf4c9deec6abc1a6b2e8984