change decl_hash to reflect deregister_php_stdlib
Summary:
Our .hhconfig has `deregister_php_stdlib = true`. This means that when we read HHI files, any function or class or member with `__PHPStdLib` gets filtered out. Naturally, this alters (1) which decls there are in a file, (2) it alters the decl hash of those types that had __PHPStdLib members.

Currently the decl_hash of a decl is computed PRIOR to its members being filtered. (Decl-hashes appear in the naming table, and are used by the old decl service and by the re-architecture.) What does this mean? -- If someone used the decl-hash of a filtered hhi file, they're okay so long as they consistently use the decl-hash that they got from ParsedFileWithHashes or FileSummary. But if they computed hash themselves from the decl in their hand, they'd silently get wrong behavior.

This diff changes to now use decl-hashes AFTER hhi filtering. This will impact the content of the naming-table, the old behavior decl store, sub1m-tw, and re-architecture. But I think they're all okay low-risk changes.
* The decl-store is the only one with production-facing impact, but HHI files never change, so it will never even arise... Reason is that the old decl service is called in the event of a file-on-disk being changed (to fetch the previous version of the file's decls, so as to do fine-grained decl-diffing). The Hhi files necessarily won't ever be changed, so we'll never query the old decl service for them.
* (sub1m TW does query for old decls to assemble its inputs, but that's not a production scenario, and in any case sub1m TW definitely has other bugs in it relating to deregister_php_lib)

Reviewed By: edwinsmith

Differential Revision: D41448381

fbshipit-source-id: 660d5cec4964a3fb508c5322d71994677758f536