Restructure static docs
Summary:
Move the documentation pages according to the new table of content (TOC). One notable difference from the original TOC is that the structure around the Compiler section is flatter since we don't have content yet.

Pages content of which needs to be incorporated elsewhere are moved to the `_migration` directory.

Reviewed By: iahs

Differential Revision: D41495865

fbshipit-source-id: 8594b995101e7f505a9f831e0f9cf05c05051ac3