Refactor enforcement calculations
Summary:
Very similar code for determining whether a type is enforceable or not was present in Decl_enforceability and Typing_enforceability, and they were starting to diverge from each other. We refactor so that there is only one implementation. The implementation needs to be parameterised over the representation of classes, because the Decl one is used during folding and uses shallow classes (and can't register dependencies), and the Typing one is used during type checking and uses folded classes (and must register dependencies). Hence, we functorise to parameterise over the representation of classes and contexts, and supply various helper methods.

This diff is intended just to do the refactoring, subsequent diff will clean up. However, there were already some differences between the two, especially Typing considering `supportdyn<t>` enforceable iff `t` is.

Reviewed By: andrewjkennedy

Differential Revision: D41474132

fbshipit-source-id: 9ef76fe174e4f79ab894703b71e1e19d5d9ff62b