Introduce config option so HH_FIXME silences all errors regardless of code
Summary:
Add an `.hhconfig` (or `config=` on command-line) flag to `hh` named `code_agnostic_fixme`, the effect of which is to treat *any* `HH_FIXME` as silencing all errors, regardless of error code.

To be used only for experiments, primarily Sound Dynamic pessimisation that has the effect of changing error codes.

Reviewed By: madgen

Differential Revision: D41530077

fbshipit-source-id: b3ab04dae2b777f572b883216b82a4974a83b065