more type xrefs
Summary:
Fill the `hack.TypeInfo` predicate with more xrefs. Before this diff, we only generated xrefs for types of the form "class_name".

This diff generates xrefs for the most common cases (`Awaitable<?Foo>`, `vec<PAYGraphQLPaymentItemInput`...).

For that, I wrote a new pretty printer. It is still incomplete and fall backs on the "official" pretty-printer `Typing_full.decl`  for some types but I found it was the simplest approach. Other option would be to adapt `Typing_full.decl`
 for xrefs generation but probably more work.

Reviewed By: donsbot

Differential Revision: D41518424

fbshipit-source-id: f2ad93eec95ac8e9357315a9aa3b4d8a216b9b79