Typecheck `ext_reflection-internals-functions`
Summary: This file contains some functions used by reflection for modifying objects in a dynamic manner. We ... had a default in the middle of a function, which was something I didn't realize HHVM even understood. Grepping for uses of these functions indicates that they're always used with all arguments, so I'm just removing the defaults.

Reviewed By: aorenste, jano

Differential Revision: D38961369

fbshipit-source-id: 40225693c76b212ecaa3e646d7e99c11423bf9bb