I am the creator and the destroyer of rupro
Summary:
We established that rewriting the type checker in Rust is not really viable at this time (not line by line; not in a Rust-idiomatic way)

Rupro was an experiment. The experiment is over. If we'd try again for real we would start over anyways (have you looked at the horrible subtyping prototype?)

Reviewed By: enetsee, edwinsmith

Differential Revision: D40801273

fbshipit-source-id: 70c39871af04f659d7058e12bb9b6b855be1673b