Error If Accessing Undefined $this
Summary:
# Context
{D7763002 (https://github.com/facebook/hhvm/commit/40e2c9ef5e7850df3e73cd12c5b3494736c28ea3)} changed error reporting of undefined locals from occurring in the naming phase to be done during typing, as part of the continuation based typing change. This made changes to when we report local variables as undefined, but did not change how `$this` variables are handled. `$this` is handled specially and has its own AST node (`This`). We never reported an error when accessing the `$this`, even if it is undefined. Instead we would silently provide a `Tany` type.

For the most part this was ok because we have various checks that limited where `$this` would appear. Particularly for constants, since they needed to be static expressions, they did not allow variables like `$this` to be used. This changed with the introduction of enum classes that lifted these restrictions. Thus the `$this` variable would not be defined as part of an enum class expression, yet no error would be reported. Example exhibiting this would be:

```
final class Foo {}

enum class Foos: Foo {
  Foo EXAMPLE = $this;
}
```

Here `$this` appears as part of an enum class member declaration, and since it is not defined has type `Tany` and thus passes the subtype check that it is a `Foo`

# Mitigation

To fix this I changed the typing code for `This` expression to `Typing_env.get_local_check_defined`. This will properly report an error if the variable is not defined in any active continuation. This matches the handling we do for `Lvar`.

I also considered reporting this as an error earlier in the pipeline, but there isn't a great place to do this. The other place we report errors based on `$this` usage is in parsing on the full fidelity parse tree, but at this point we haven't desugared enum classes and in general would not fix this for future features we might add. This felt like the best option to prevent future errors, though a larger refactor should be considered to avoid so much special handling of the `$this` variable.

Reviewed By: nt591

Differential Revision: D41591914

fbshipit-source-id: a039fa7a4b605f4bddac224f326051a01b483a99