Use enforcement based off declared signature not inherited signature
Summary:
When checking signatures for enforcement, we should take care to use the signature from the *origin* not the inherited one. That's because HHVM enforces according to the syntax, and doesn't account for generic instantiations. For example,
```
class B<T> {
  public function get(): T { ... }
}
interface I {
  public function get(): int { ... }
}
class C implements I extends B<int> { }
```
Here, the return type of the inherited method `C::get` should be treated as enforced, as its declaring (origin) signature has a generic parameter as return type.

Differential Revision: D41530607

fbshipit-source-id: 0b0178565ae03673e519054c63af79648f6e353c