Enable Thrift to print values with gtest
Summary:
This diff introduces a function `PrintTo` able to print any thrift type.

The function is defined into its own file.
Google Test searches `PrintTo` by ADL to print values in tests.

This diff only introduces the function, without injecting it into the namespace of the types. At this point, this needs to be done manually (see the tests).

A following diff will change the compiler to automatically inject this function into the namespace.

I'm wasting so much time looking at test failures with

```
Failure
Value of: response
Expected: is an union with `success` active, which (is an object whose field `newEventState` is equal to 168-byte object <E8-EF 25-00 F0-60 00-00 00-00 00-00 00-00 00-00 00-BE BE-BE BE-BE BE-BE BE-BE BE-BE BE-BE BE-BE 08-F0 25-00 F0-60 00-00 00-00 00-00 00-00 00-00 00-BE BE-BE BE-BE BE-BE BE-BE BE-BE BE-BE BE-BE ... 00-00 00-00 00-00 00-00 BE-BE BE-BE BE-BE BE-BE 00-00 00-00 BE-BE BE-BE 90-00 75-00 B0-60 00-00 90-00 75-00 B0-60 00-00 90-00 75-00 B0-60 00-00 01-00 00-00 00-00 00-00 00-00 00-00 00-01 BE-BE>) and (is an object whose field `diffFromStartingState` is equal to 56-byte object <BE-BE BE-BE BE-BE BE-BE 00-00 00-00 BE-BE BE-BE C0-F0 25-00 F0-60 00-00 C0-F0 25-00 F0-60 00-00 C0-F0 25-00 F0-60 00-00 01-00 00-00 00-00 00-00 01-BE BE-BE BE-BE BE-BE>)
  Actual: 264-byte object <68-96 21-00 F0-61 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 88-96 21-00 F0-61 00-00 00-00 00-00 00-00 00-00 00-40 E4-E9 44-7F 00-00 48-41 E4-E9 44-7F 00-00 ... 70-E6 25-00 F0-60 00-00 01-00 00-00 00-00 00-00 00-BE BE-BE BE-BE BE-BE 20-2F 85-00 80-60 00-00 78-2F 85-00 80-60 00-00 78-2F 85-00 80-60 00-00 01-01 01-BE BE-BE BE-BE 01-00 00-00 BE-BE BE-BE> (of type facebook::monitoring::events::event_store::decoration::EventDecorationStatus), whose active member `success` is 256-byte object <68-96 21-00 F0-61 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 88-96 21-00 F0-61 00-00 00-00 00-00 00-00 00-00 00-40 E4-E9 44-7F 00-00 48-41 E4-E9 44-7F 00-00 ... 70-E6 25-00 F0-60 00-00 70-E6 25-00 F0-60 00-00 01-00 00-00 00-00 00-00 00-BE BE-BE BE-BE BE-BE 20-2F 85-00 80-60 00-00 78-2F 85-00 80-60 00-00 78-2F 85-00 80-60 00-00 01-01 01-BE BE-BE BE-BE> (of type facebook::monitoring::events::event_store::decoration::EventDecorationSuccess), whose field `newEventState` is 168-byte object <68-96 21-00 F0-61 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 88-96 21-00 F0-61 00-00 00-00 00-00 00-00 00-00 00-40 E4-E9 44-7F 00-00 48-41 E4-E9 44-7F 00-00 ... 00-00 00-00 00-00 00-00 BE-BE BE-BE BE-BE BE-BE 00-00 00-00 BE-BE BE-BE 00-00 00-00 00-00 00-00 D0-96 21-00 F0-61 00-00 D0-96 21-00 F0-61 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-01 BE-BE> (of type facebook::monitoring::events::event_store::decoration::EventState)
```

This diff also moves the files into a `util/gtest/` folder, so that it's clear that this is part of the library and is intended to be used by users, and it's not for internal testing.

Reviewed By: iahs, vitaut

Differential Revision: D40270486

fbshipit-source-id: e0bfd449808121f99adcf3821994287594162571