Speed up hhdgdiff comparisons
Summary:
hhdgdiff walks every hash in one hash list, and looks it up using
`DepGraph::dependent_dependency_edge_exists` in the other `DepGraph`:

```
            if let Some(list) = dg1.hash_list_for(dependency) {
                for dependent in dg1.hash_list_hashes(list) {
                    if !dg2.dependent_dependency_edge_exists(dependent, dependency) {
```

This is slow because:

1. it does a big binary search in the inner loop to look up the same HashList every time
2. within that HashList it needs to do a binary search to check membership

Instead, just fetch each `HashList` once and use a `HashSet` to check membership in O(1) per edge.

With some minor tweaking, this could check both for missing edges and extra edges, by doing a
`HashSet::remove` to check membership, and then see if the `HashSet` is empty at the end.
But this wouldn't show the edges for `Dep`s that are completely absent in the second `DepGraph`
as the current code does so I didn't do this. Those will be reported separately so maybe
we don't care?

Reviewed By: edwinsmith

Differential Revision: D41614484

fbshipit-source-id: 16bfbf67775cd927c7cab1848b5426b6eec1f8fb