Add support for buck2 locations when using hhvm_wrapper.php
Summary:
`hhvm_wrapper.php` finds the HHVM binary built by buck. As we move to buck2, we should search for it there too (and probably eventually get rid of the old buck search).

This also removes the search in `_bin`, which doesn't look to be needed anymore.
Also removes the `--build-root` option; we should be able to just use `--bin`, especially since `hphp` is invoked via `hhvm` and this helper function is specifically for find hhvm anyway. Please let me know if this is still in use by anyone.

Reviewed By: jtwarren

Differential Revision: D41629914

fbshipit-source-id: c18c057a4bfae1835e1138fd5df3d2e4bb18dc31