Default to `&` for return type of `coro::transform`
Summary:
`AsyncGenerator` by default expects a reference as first argument.
`transform` instead makes it very easy to just return a value.

The typical transformer is going to be a lambda, which typically returns by value.
But what the lambda returns should necessarily be what the dereference of the generator result should return.

This diff introduces 2 changes:
1. By default, the return of the AsyncGenerator is a reference. This avoids copying the value each time the iterator is dereferenced
2. Make it practically usable to specify the type of the returned generator. Right now it's possible to override the return type of the generator, but it's close to impossible to use, as a user first need to define all the template arguments, including the function used for transforming. This makes it unusable if someone wants to use an inline lambda.

The implementation forwards the reference directly if it's compatible, otherwise it creates a copy and forwards a reference to that.
See https://godbolt.org/z/vErzPW4P3 for the compatibilities.
The lines that don't compile in that example are the ones where a copy is create explicitly in this implementation.

NOTE: in some of the parts that work, the reference is constructible because a temporary is going to be created. So a copy is already performed there, it's just implicit.
That is why I decided to support the behaviour and create a copy explicitly, since it's not more expensive than other cases.

Reviewed By: Gownta

Differential Revision: D41399318

fbshipit-source-id: 0d2759859a1878ddf52e776397e6d17dabddcea4