Remove more problematic definitions from the types library
Summary:
`SourceLocation` and `SourceRange` are a weird mix of resolved and unresolved source location/range representation. They are unused, don't belong in the types library and superseded by `ast.SourceRange`. Remove them in favor of `ast.SourceRange`.

Similarly, `UniversalHashAlgorithmEnum`, `minTypeHashBytes` and `defaultTypeHashBytes` are unused except for the conformance test and the last two expose implementation details that don't belong to the types library as it is currently described. Remove them too.

Reviewed By: iahs

Differential Revision: D41605054

fbshipit-source-id: fd5a32cd926fd64b17d35bc745cbb0f8fd0bec6d