Type SDT closures by making captured variables like types and skip typing on SDT passes
Summary:
See comment in `typing.ml` for illustration.

Currently, we type closures both in normal mode and SDT mode, which means that we will type a closure at 2^depth times because we repeat each pass for the dynamic pass of its outer function. This is really slow.

In HackC, we compile closures to classes with `__invoke` methods where captured variables are private properties. Naively, we make classes SDT by making local variables be like types.

This diff simply makes captured variables be like types when typing an SDT closure. When we encounter a closure on an SDT pass, we skip typing it entirely (throw out the node and replace with `Tast.Omitted`), and just give it a `dynamic` type. This way, we always type a closure a maximum of 2 times regardless of depth at the cost of a little precision.

Differential Revision: D41641174

fbshipit-source-id: 1c10194c6797586c0017eee77a2d9daf3ae4c389