Remove support for populate_member_heaps; simplify oldify/get_old/remove_old/remove
Summary: We use `populate_member_heaps` in 100% of prod use cases except Zoncolan, and I think they intend to migrate to it. If we need to re-add support in order to roll out, we can (without changing these FFIs--when we get a request to oldify/remove a shallow class, we can look it up from the shallow classes store on the Rust side and remove all its members from the member stores). But I think it's better to just remove support for now, since it lets us simplify things quite a bit.

Reviewed By: shayne-fletcher

Differential Revision: D40813624

fbshipit-source-id: c5baa655c72117a226aa2cd2c715d8902e9a58d5