private properties in a class and a trait that requires the class are aliased, not shadowed
Summary:
Whenever a trait and class that uses the trait both define a private property, the two definitions are aliased.  In presence of `require class` constraints, as in `aliased_private_properties_01.good.php` Hack was incorrectly reporting:

```
Typing[4256] Typing error [1]
-> This object access to i is ambiguous [2]
-> You will access the private instance declared in T [3]
-> Instead of the private instance declared in D [4]
```

This diff makes the check if private properties are shadowed or aliased aware of the `require class` case.

Reviewed By: vassilmladenov

Differential Revision: D41581171

fbshipit-source-id: 4d7be33f6d8f04c03a1d8b88c73201451f5a9232