Typing support for loose refinements
Summary:
This diff implements refinements in their full
generality. The core associated subtyping
rule is obvious:

```
L2 <: L1      U1 <: U2
------------------------------------------------
{type T super L1 as U1} <: {type T super L2 as U2}
```

The devil is in the details, as usual. Tricky
logic is required to interact reasonably well with
type accesses. In Typing_subtype_tconst we raise
errors when a type access is performed on an
inexact refinement; this is tested in `accessloose.php`.
In Typing_taccess we use refinements to get
more precise bounds on the result of type
accesses. However, when an access is about
to happen on a loose type refinement, we raise
an error. One test case in `accesslose.php` is
supposed to verify that the error is indeed
flagged, but it is currently not working because
localization code sometimes drop errors.

Some effort was put into generating good
errors. This involves a couple tricks, like
useless subtype calls just here to catch errors
early; some comments explain how this all
works.

Reviewed By: losvald

Differential Revision: D38861371

fbshipit-source-id: 212f59e7fdbe40f48999d64cbb5e2496634f8074