Use the literal array/static string tables for all serialization of UnitEmitter
Summary:
Right now we use the literal array table and static string table to
encode arrays and strings in the bytecode for an UnitEmitter. However,
many strings and arrays are stored in the UnitEmitter outside of the
bytecode. Modify the serialization logic of UnitEmitter to use these
tables for arrays and strings anywhere within the unit.

Reviewed By: edwinsmith

Differential Revision: D41715186

fbshipit-source-id: ddcf32add41f777c64f48d70ea1b109343c46693