Multifile parsing: don't add a blank line at the end of files
Summary:
To split the multifile properly,  the delimiter should have a newline character both before and after, otherwise it will add a blank line at the end of files. For such a delimiter to work we also need to add a newline char at the beginning of the file content.

This will matter for fanout tests where properly detecting file changes is important and we don't want random new lines to be detected.

Reviewed By: andrewjkennedy

Differential Revision: D41659493

fbshipit-source-id: 145b711de85772f42f8883ffd2c5483f59f52e09