report same folding edges as OCaml decl lib
Summary:
hh_distc uses hackrs to do decl folding, and relies on FoldedDeclProvider reporting dependency edges for folded decls. These fixes were necessary to get to parity with `hh` folding dep edges (from decl/decl_env.ml and decl_folded_class.ml)

1. each class -> parent relationship gets one Type and one Extends edge,
but only if the class is not an Hhi decl.
2. each class gets a Type edge to every ancestor, whether or not
the class or the ancestor is an Hhi decl.

The hhi filtering is weirdly inconsistent - direct parents are filtered out
but indirect parents are not.

Open questions
* does the rust hh_server backend (hackrs_provider_backend) solve this some
other way to have parity with hh_server's dep-edges?
* should I refactor this logic so only hh_distc uses it?
* why does hh_server filter hhi edges at all? (they are a tiny fraction of all edges)
* why doesn't hh_server filter hhi edges consistently?

Differential Revision: D41531974

fbshipit-source-id: 46d3bd121e740c186c973c78ca643311d4576aeb