Separate `class_id` elaboration
Summary:
Currently, the lowerer will emit `CIexpr(...)` which are converted to `CIparent`, `CIself`, `CIstatic`, `CI`, `CIexpr(_,_,Lvar _)` and `CIexpr(_,_,This)`. This is interwined with validation of `Class_get`, `Class_const`, `FunctionPointer` and `New` expressions and of statements and expressions more generally.

This diff seperates the `class_id` elaboration pass into a visitor.

Reviewed By: CatherineGasnier

Differential Revision: D41067076

fbshipit-source-id: b425a03f065d012280327bc19c268e70e3fb5725