Elaborate `Lvar` expression
Summary:
The lowerer gives us an AST where all local variable are represented by `Lvar _`. This diff elaborates `$$` , `$this` and `$_` to their canonical representations i.e. `Dollardollar`, `This` and `Lplaceholder`.

The current naming pass does not perform this elaboration consistently. In particular:
- `Lvar _` appearing as the key expression in `As_kv` would not be elaborated to `Lplaceholder`
- `Lvar $$` appearing in a class get expression would not be elaborated to `Dollardollar`

There are ~~two~~ three places in typing which rely on this oversight, `bind_as_expr` and `Nast.class_id_to_str`, used in `typing_fake_members` and ifc. These are fixed in this diff.

Reviewed By: CatherineGasnier

Differential Revision: D41306284

fbshipit-source-id: 8068a509046e3f5b42be77fa8eb12a2a65cc3ff5