delete dump_saved_state_depgraph.cpp
Summary:
This C++ program reads a sqlite3 format depgraph and prints it.

I admit I'm not fully aware of all the depgraph formats in flight, particularly
which are new for the rearchitecture and which are legacy tech debt.
I'm guessing sqlite3 depgraphs are no longer in use, and offering this cleanup
diff as a public service.

Is there other dead code for reading or writing sqlite3 dep graphs we should
delete as well? Or is this format still alive and kicking?

Reviewed By: ljw1004

Differential Revision: D41578428

fbshipit-source-id: 3a195e1b86743576da336efa7f52ac725b76d02c