Do not include classname etc fields in type structures off unbounded type constants
Summary: If a type constant has no bounds, then we unsoundly assume that a type structure built from it has `classname` and other fields that are only relevant for types that are classish. But give it a `mixed` bound (which should be equivalent to no bound), and we don't make this assumption. Let's fix this anomaly (discovered when pessimising using bound of `supportdyn<mixed>`).

Reviewed By: vassilmladenov

Differential Revision: D40260697

fbshipit-source-id: 4ad99929a08a606976e6ab12cfc2dd78bb52b435