Use module private reference in `naming_captures`
Summary:
Currently we wrap our visitor in a function and construct a new visitor object on each invocation of `Naming_captures.elab`. This lets us hide a reference within the function with which we accumulate captured local vars.

This diff moves that reference outside of the function and brackets the `elab` call to reset the module private state between functions. This means we define the visitor object only once.

Reviewed By: CatherineGasnier

Differential Revision: D41717218

fbshipit-source-id: 2083392b2310618fea5d4048ed0e972b5a9e9ff9