Eliminate explicit continuation
Summary: As with the original definition, the naming passes are sequenced so that we never require the transformation to be restarted after some AST element is transformed. This meant we could simplify the `Cont.t` continuation type and `Transform.combine` function. We then have only a `Next` and `Finish` continuation making the type isomorphic to `Result.t`. There is already a short-circuiting fold over `Result.t list` in Core so we redefine our fusion in terms of that function.

Reviewed By: CatherineGasnier

Differential Revision: D41614431

fbshipit-source-id: f8748de901b710ad28a3c7148eea36dfd01fd4ed