Move all mapreduce visitors to normal form
Summary:
The derived visitors follows our types. Whenever we have a type definition of alias, it will create a corresponding method.

However, during elaboration / validation, we often want to:
(1) operate on some _part_ of a type, e.g. a record field
(2) operate on a list of some type
(3) distinguish the use of a type in certain positions

Note that we could achieve (2) and (3) by modifying the AST definition, adding more type aliases but this doesn't help with (1).

Providing these methods helps to reduce the size of the visitor code (since we can recmove all boilerplate) and allows us to put the visitor into 'normal form'.

Normal form means that we modify either the AST element, the environment, generate an error or some combination of the three and then either return the element _traverse_ the element to transform its sub-elements. This will later allow us to phrase each traversal as a fusable transformation.

This diff defines a custom mapreduce class which adds the specific methods need and modifies each mapreduce visitor to inherit from it and express its transformation in normal form

Reviewed By: CatherineGasnier

Differential Revision: D41469100

fbshipit-source-id: a4d6792f7ea75ef7888f10e7a17c01325d38b7f8