Rephrase visitors as composable passes
Summary:
Since each pass is now defined as its own visitor, we are doing considerably more traversals of the AST than in the original naming phase.

This diff rephrases our pass as a _composable_ pass that will be fused in a subsequent diff. This gives us the benefit of defining our passes separately without incurring the perf hit.

Reviewed By: Wilfred, shayne-fletcher

Differential Revision: D41581658

fbshipit-source-id: aa40649d68dfdfdfa009da4f80145f37f45f8044