Optimize traversal
Summary:
In the original naming phase, there was a hand written traversal of the AST with manually fused transformation.

In this stack we have been carefully separating each transformation before automatically fusing, using a ppx derived visitor for the traversal.

The derived visitor has the benefit of... being derived *but*, and it is a **big** but, it turns out it's costing us ~5% on full init and fanout heavy scenarios.

Why?

First it is because the visitor spends a lot of time traversing AST elements that we never want to modify and contain no child elements we want to modify. In short, we spend a lot of time traversing parts of the AST for no reason.

Second, it is because there is a performance overhead to using objects in OCaml: https://discuss.ocaml.org/t/actual-performance-costs-of-oop-objects/9400/2

> My rule of thumb is that if objects are used heavily, you are going to see some cost (to be weighted against the benefits of this programming style, in particular the extensibility/openness of row types), otherwise there is probably nothing to worry about (just look at the profiles).

> For example: **the `visitors` package uses objects to implement AST traversals. If visitors-derived traversal on large objects are the critical path of your application, you will probably observe object performance**. If you just use it in passing, it’s fine. For example if we used visitors in the compiler, it would probably be fine, because structural mapping/folding is only a fraction of the work we do. On the other side, the `Gilian` project had a large object-related performance regressions because it uses this kind of objectful AST traversals in performanc-critical paths. (To implement formula substitutions in a symbolic analyser, if I remember correctly.)

This diff removes the derived visitor in favor of a function based traversal where we support fusing by providing a record of functions to transform each element.

This completely eliminates the regression we see earlier in the stack.

Reviewed By: CatherineGasnier

Differential Revision: D41717247

fbshipit-source-id: 318860fd7ce5ad61222c1aea7ea7549ef9e94710