rename decl_hash -> file_decls_hash in naming_sqlite.ml
Summary:
This diff is a pure rename. It's because I found it too confusing to see the identifier name `decl_hash` used for both the FileDeclsHash (stored once per FileInfo.t, represents the hash of the file as a whole) and also for the DeclHash (stored once per decl, represents the hash of just that decl).

I picked the name `file_decls_hash` since that's what it's currently called in Rust.

Reviewed By: bobrenjc93

Differential Revision: D41438027

fbshipit-source-id: 69dbfa118bd573bda404d3065eee950e25f7f158