remove prepare_for_typecheck
Summary:
This API no longer does anything.

It had been introduced for the RPC decl process, as a mitigation against the fact that sometimes a decl would no longer be pinned while we're typechecking a big file.

I haven't yet observed evidence of needing it with hh_decl_shmem. If it is needed, I can bring it back.

I think that as well as just perf, it used to have correctness implications. Consider the case when you're typechecking an unattached file "/tmp/a.php" with respect to a repository "/data/www/.hhconfig". When folks used this prepare_for_typechecking, then all the top-level-decls from a.php would become available when you typecheck other files, even though they might not be the winner. I think this might result in folding bugs (e.g. what if a class here is folded using a shallow class from here, but another class elsewhere is folded using a differnet shallow class).

Anyway, I don't want to support that scenario, and in D41252121 I removed it for the only (test) user of the decl-service.

Reviewed By: zhyty

Differential Revision: D41254652

fbshipit-source-id: 1795a23c4aa252ddbaceadea8203888267ad455f