hh_single_type_check --get-some-file-deps
Summary:
This new "mode" prints a list of files the argument file depends on. It takes integer which represents the depth of the traversal.
This is to satisfy Infer's development need. Not all dependencies are fetched, but a set of relevant files is returned for Infer to perform useful analysis.

Reviewed By: ljw1004

Differential Revision: D41767578

fbshipit-source-id: 2364dfeed498f12a402ae036554390d270e2a884