Check for larger number of bytes before filling objectdata region with debug characters
Summary:
D40482293 (https://github.com/facebook/hhvm/commit/0ce9142204109f7d9e78a509ba0c432dd6c3a3a8) added an object's header and class to the bytes immediately following the FreeNode, when freeing the object, to aid in debugging.

The [two smallest size classes](https://www.internalfb.com/code/fbsource/[2285653cad61631e128a5e4869583a8bc5571e9d]/fbcode/hphp/runtime/base/memory-manager-inl.h?lines=264) are 16 and 32, and so **things are fine as is**.

However, there's a risk when `bytes` >= 16 and < 32 (say when it's 24); this currently doesn't happen because `debugFreeFill` is only called with a number of bytes coming from `SIZE_CLASSES` table. But if that ever does become possible (we've already compacted the object properties to be unaligned, so a 24-byte object isn't unfeasible if we decide to change the `SIZE_CLASSES` table), the code as is would fail. This is because it guards against the size of HeapObject, but uses the sizeof FreeNode to do the filling (so theoretically a negative `count` argument to `memset` would be possible).

I think we should instead check against the size of FreeNode before doing proceeding, in particular because `debugFreeFill` doesn't necessarily need to be called with a value from the `SIZE_CLASSES` table/it accepts arbitrary values for `bytes`.

Reviewed By: binliu19

Differential Revision: D41754675

fbshipit-source-id: 931314c5e1bb40da65f7c77580587c1c9c04b2fa